requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cloudinary==1.21.1
Django==3.0.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
djangorestframework==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
